chore(analytics): lower case providers directory for consistency
