+++ packages/analytics/__tests__/Analytics.test.ts
@@ -9,7 +9,7 @@ jest.mock('../src/vendor/dom-utils', () => {
 import { ClientDevice, parseAWSExports, Hub } from '@aws-amplify/core';
 import { AnalyticsClass as Analytics } from '../src/Analytics';
 import { AnalyticsProvider, PromiseHandlers } from '../src/types';
-import { AWSPinpointProvider as AWSAnalyticsProvider } from '../src/Providers/AWSPinpointProvider';
+import { AWSPinpointProvider as AWSAnalyticsProvider } from '../src/providers/AWSPinpointProvider';
 
 jest.mock('@aws-amplify/core');
 const mockHubDispatch = Hub.dispatch as jest.Mock;
