+++ packages/analytics/__tests__/providers/EventBuffer.test.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import EventBuffer from '../../src/Providers/EventBuffer';
+import EventBuffer from '../../src/providers/EventBuffer';
 
 const DEFAULT_CONFIG = {
 	bufferSize: 1000,
