+++ packages/analytics/pinpoint/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@aws-amplify/analytics/pinpoint",
-  "main": "../lib/Providers/pinpoint/index.js",
-  "browser": "../lib-esm/Providers/pinpoint/index.js",
-  "module": "../lib-esm/Providers/pinpoint/index.js",
-  "typings": "../lib-esm/Providers/pinpoint/index.d.ts"
+  "main": "../lib/providers/pinpoint/index.js",
+  "browser": "../lib-esm/providers/pinpoint/index.js",
+  "module": "../lib-esm/providers/pinpoint/index.js",
+  "typings": "../lib-esm/providers/pinpoint/index.d.ts"
 }
