+++ packages/analytics/src/index.ts
@@ -4,9 +4,7 @@
 // TODO(v6) Remove once all default provider functional APIs available
 export { Analytics } from './Analytics';
 export { AnalyticsProvider } from './types';
-export {
-	AWSPinpointProvider,
-} from './Providers';
+export { AWSPinpointProvider } from './providers';
 
 // TODO(v6) Refactor as additional Analytics providers come online
 /*
@@ -14,8 +12,8 @@ export {
 	AWSKinesisProvider,
 	AWSKinesisFirehoseProvider,
 	AmazonPersonalizeProvider,
-} from './Providers';
+} from './providers';
 */
 
 // Default provider types
-export * from './Providers/pinpoint';
+export * from './providers/pinpoint';
