+++ packages/analytics/src/providers/AWSKinesisFirehoseProvider.ts
