+++ packages/analytics/src/providers/AWSKinesisProvider.ts
