+++ packages/analytics/src/providers/AWSPinpointProvider.ts
