+++ packages/analytics/src/providers/AmazonPersonalizeHelper/SessionInfoManager.ts
