+++ packages/analytics/src/providers/AmazonPersonalizeProvider.ts
