+++ packages/analytics/src/providers/pinpoint/index.ts
