+++ packages/analytics/src/types/Analytics.ts
@@ -83,5 +83,5 @@ export interface AnalyticsEvent {
 	immediate?: boolean;
 }
 
-export { PersonalizeAnalyticsEvent } from './Providers/AmazonPersonalizeProvider';
-export { KinesisAnalyticsEvent } from './Providers/AWSKinesisProvider';
+export { PersonalizeAnalyticsEvent } from './providers/AmazonPersonalizeProvider';
+export { KinesisAnalyticsEvent } from './providers/AWSKinesisProvider';
