+++ packages/analytics/src/types/index.ts
@@ -3,4 +3,4 @@
 
 export * from './Analytics';
 export * from './Provider';
-export * from './Providers';
+export * from './providers';
